requirements.txt
pandas==1.4.1
Jinja2==3.0.3
Django==4.0.4
requests==2.27.0
channels==3.0.4
channels_redis==3.4.0
daphne==3.0.2
whitenoise==5.1.0
paho-mqtt==1.5.0
numpy==1.22.1
fpdf==1.7.2
pytz==2021.3
xlrd==2.0.1
django-admin-shell==1.0.0